Further subdivision
Further divided functions up into separate modules
